# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API reto Torneo Solidario",
        "description": "API reto Torneo Solidario",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/equipos": {
            "get": {
                "tags": [
                    "equipos"
                ],
                "summary": "Obtener equipos",
                "description": "Obtener todos los equipos",
                "operationId": "getEquipos",
                "responses": {
                    "200": {
                        "description": "Equipos obtenidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/equipos"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "equipos"
                ],
                "summary": "Crear un equipo",
                "description": "Crear un nuevo equipo",
                "operationId": "postEquipo",
                "requestBody": {
                    "description": "Datos del equipo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "centro_id",
                                    "grupo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Equipo 1"
                                    },
                                    "centro_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "grupo": {
                                        "type": "string",
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Equipo creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/equipos"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no válidos"
                    }
                }
            }
        },
        "/api/equipos/{id}": {
            "get": {
                "tags": [
                    "equipos"
                ],
                "summary": "Obtener un equipo",
                "description": "Obtener un equipo por su id",
                "operationId": "getEquipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/equipos"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "equipos"
                ],
                "summary": "Actualizar un equipo",
                "description": "Actualizar un equipo por su id",
                "operationId": "updateEquipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del equipo",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "centro_id",
                                    "grupo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Equipo 1"
                                    },
                                    "centro_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "grupo": {
                                        "type": "string",
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Equipo actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/equipos"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "equipos"
                ],
                "summary": "Eliminar un equipo",
                "description": "Eliminar un equipo por su id",
                "operationId": "deleteEquipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del equipo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Equipo eliminado"
                    },
                    "404": {
                        "description": "Equipo no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "equipos": {
                "title": "equipos",
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Equipo 1"
                    },
                    "centro_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "grupo": {
                        "type": "string",
                        "example": "A"
                    },
                    "usuarioIdCreacion": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fechaCreacion": {
                        "type": "timestamp",
                        "example": "2022-02-11 15:12:24"
                    },
                    "usuarioIdActualizacion": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fechaActualizacion": {
                        "type": "timestamp",
                        "example": "2022-02-11 15:12:24"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}